Fixed broken links in footer (on all pages)
All of the links in the footer of every page were broken (they linked to nowhere).

In order to resolve this I added ids to sections that corresponded with the navigation links and pointed the href attribute to the corresponding section id. For the links to external websites, I updated the href value to contain the correct URL.